requirements.txt
torch>=2.0.0
transformers>=4.41.0
huggingface_hub>=0.16.0
accelerate>=0.31.0
safetensors>=0.4.2
numpy>=1.20.0
tqdm>=4.65.0
python-dotenv>=1.0.0
